Rewrite client structure with NextJS, and add basic site functionality
The project client-side has been rewritten from scratch (the html was
copied) to use NextJS as it's frontend, as it would have otherwise required an
numerous time invested into writing something that could've been spent
elsewhere.

I have added the basic functionality of choosing a username and a
demo version of the contacts list.