Improve site functionality & security
The characters that can be used on a username have been cut down to A-z,
digits, underscore, and dashes - both checks have been implemented in
the server & client sides.

A classic issue with the upper/lower cases usernames has been fixed. The
usernames also get corrected on the server-side if there's a mismatch
with the client side's input.

An issue with the messages & contacts list not being automatically
scrolled down, that has been fixed.

There was frankly, a lazy decision I made with the ContactResponse
packet. It was being used incorrectly - That has been migrated by adding
a new packet called ContactAdd packet.